+++ tfjs-core/src/ops/ops.ts
@@ -46,6 +46,7 @@ export {divNoNan} from './div_no_nan';
 export {dot} from './dot';
 export {equal} from './equal';
 export {eye} from './eye';
+export {fill} from './fill';
 export {greater} from './greater';
 export {greaterEqual} from './greater_equal';
 export {less} from './less';
